put some functions that were misplaced in correct place and cleaned up code
