+++ app/models/Users.java
@@ -33,36 +33,6 @@ public class Users {
      * @param isMod - user's moderator status
      * @return - boolean validating the success of adding a user to the User class database
      */
-    public static boolean addUser(String name, String password, String email, boolean isMod) {
 
-        return true;
-    }
 
-    /**
-     * verifies that a user exists in the User class database
-     * @param name - a user's name
-     * @param password - a user's password
-     * @return - boolean validating whether or not a user exists
-     */
-    public static boolean verifyUser(String name, String password){
-        return true;
-    }
-
-    /**
-     * removes a user from the User class database
-     * @param id - a user's id
-     * @return - Result validating whether or not a user has been removed from the User class database
-     */
-    public static Result deleteUser(int id) {
-
-        User user = User.find.select("id").where().eq("id",id).findUnique();
-
-        if(user == null) {
-            return notFound();
-        } else {
-            user.delete();
-            return ok();
-        }
-
-    }
 }
