+++ conf/application.conf
@@ -32,28 +32,27 @@ application.langs="en"
 # ~~~~~
 # You can declare as many datasources as you want.
 # By convention, the default datasource is named `default`
-#
 #db.default.driver=org.h2.Driver
 #db.default.url="jdbc:h2:mem:play"
 # db.default.user=sa
 # db.default.password=""
 #
 
-#db.default.driver=com.mysql.jdbc.Driver
-#db.default.url="jdbc:mysql://localhost:3306/hvz"
-#db.default.user="root"
-#db.default.password="HiToby"
+db.default.driver=com.mysql.jdbc.Driver
+db.default.url="jdbc:mysql://localhost:3306/hvz"
+db.default.user="root"
+db.default.password="HiToby"
           #HiToby is set as password for Beth's MySQL root user
           #localhost:3306/hvz  is Beth's current setup for the db
 
 # You can expose this datasource via JNDI if needed (Useful for JPA)
-# db.default.jndiName=DefaultDS
+ #db.default.jndiName=DefaultDS
 
 # Evolutions
 # ~~~~~
 # You can disable evolutions if needed
 # evolutionplugin=disabled
-#applyEvolutions.default=true
+applyEvolutions.default=true
 
 
 # Ebean configuration
@@ -61,7 +60,7 @@ application.langs="en"
 # You can declare as many Ebean servers as you want.
 # By convention, the default server is named `default`
 #
-#ebean.default="models.*"
+ebean.default="models.*"
